third_party/WebKit/LayoutTests/editing/pasteboard/drag-and-drop-objectimage-contenteditable-expected.txt

This test verifies dropping an image does nothing. To test manually, drag the image and drop it in the rect with a blue border.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageIsInserted is false
PASS successfullyParsed is true

TEST COMPLETE
